LayoutTests/http/tests/security/mixedContent/insecure-download-redirects-to-basic-auth-secure-download.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/ui-helper.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    testRunner.setHandlesAuthenticationChallenges(true);
    testRunner.setAuthenticationUsername("testUser");
    testRunner.setAuthenticationPassword("testPassword");
    if (testRunner.isWebKit2)
        testRunner.setShouldLogDownloadCallbacks(true);
}
</script>
</head>
<body>
<p>
<p>This test loads an insecure resource that redirects to a secure PDF guarded by basic authentication. The secure PDF should not be downloaded because it requires credentials and was loaded via an insecure redirect. This test PASSED if you see a JavaScript alert with message "Unauthorized". Otherwise, it FAILED.</p>
<a id="link" href="http://127.0.0.1:8080/resources/redirect.php?url=https://localhost:8443/security/mixedContent/resources/subresource/protected-pdf.php">the link</a>.
<script>
function runTest()
{
    var link = document.getElementById("link");
    UIHelper.activateAt(link.offsetLeft + 5, link.offsetTop + 5);
}
runTest();
</script>
</body>
</html>
